+++ build_vsix.py
@@ -11,7 +11,8 @@ FILES = ["package.json", "extension.js", "pdfgen.js", "svggen.js", "README.md",
          "CHANGELOG.md", "LICENSE.txt",
          "analyzer/analyze.py",
          "media/diagram.js", "media/diagram.css", "media/icon.svg",
-         "images/icon.png", "images/hero.png", "images/docstrings.png"]
+         "images/icon.png", "images/hero.png",
+         "images/filtering.gif", "images/docstrings.gif", "images/focus.gif"]
 
 MANIFEST = f"""<?xml version="1.0" encoding="utf-8"?>
 <PackageManifest Version="2.0.0" xmlns="http://schemas.microsoft.com/developer/vsx-schema/2011" xmlns:d="http://schemas.microsoft.com/developer/vsx-schema-design/2011">
@@ -56,6 +57,7 @@ CONTENT_TYPES = """<?xml version="1.0" encoding="utf-8"?>
   <Default Extension=".css" ContentType="text/css" />
   <Default Extension=".svg" ContentType="image/svg+xml" />
   <Default Extension=".png" ContentType="image/png" />
+  <Default Extension=".gif" ContentType="image/gif" />
   <Default Extension=".txt" ContentType="text/plain" />
 </Types>
 """